Create december_03.adb
Only made one mistake today, I initially stored the count of any duplicated item types for each compartment. I knew this was not required for part 1 but thought it desirable to preserve all the information in the input. I left the code in until I solved part two and later removed it to clean up the solution.

The solution verifies that both compartments include the same number of items, that is, the sum of items in both compartments is even.
Further it verifies the set of Common items has only one element, applies to both parts.