Solution to Day 7
I used this as a chance to learn how to use the Ada.Containers.Multiway_Trees.
It is the trickiest of the containers I have used. The biggest mistake was searching sub trees when executing cd xx, searching children is what is required.. The recursive calculation of directory space used is reasonably elegant.