Day 09 Solution and removal of unused packages in day 08
The solution presented has been significantly cleaned up (generalized). Part one was solved using a state variable that was a record containing the coordinates of the Head and Tail, with Head movements of more than one unit before evaluating the Tail coordinates. For part two it became clear that it was necessary to evaluate all knots after each one unit movement of the Head.