Solution Day 11
Got the first part first run. Part Two caused numeric overflow. 128 bit arithmetic was insufficient. The key is using a modulus which is the LCM of all the all the Test : numbers (all prime), that is, their product. At each calculation old is reduced to mod  LCM, doing this allows 64 bit arithmetic to be used.