Day 21 Solution
Part two is an analytical solution that finds the path to the unknown value (To_Find). There is an assumption that there is only one instance and hence it only occurs in one sub tree below root. This is a vague similar problem to one in 2017 which required the "weight" of a tree to be balanced.