Day 20 Solution
I made a right meal  of this one, my initial approach to part one was to avoid the use of access types (pointers) but that got too complex. The more obvious doubly linked ring with a map of the starting location proved much simpler to manage.
For part two I saw immediately that skipping by very large numbers was infeasible and that taking the mod based on the length of the list was the obvious solution. I could not match the example so ultimately I went to Redit and found that required modulus was the list length - 1. I think it is due to the unspecified behavior when skipping more than the length of the list. Effectively a number does not skip itself. Alternately it could be seen as the number being unlinked from the ring before the skipping starts!