Day !3 part 1 solution
Tried a completely new approach using lists! More correctly access to a variant record that is either a value or a list. Part 2 looks straight forward in that app the messages can be put into a Vector container and a generic sort applied using the ordering function from part 1.
This is the solution. to part 1.