Day 13 part two solved
Part two is relatively straightforward after the comparison function is created in part one. This allows < and = to be defined for the Packets. The generic sort on a Vector container provides the sorting procedure. Creating the  [[2]] and [[6]] list is relatively simple but tedious. General cleanup and improvement of identifier names.