Day 16 both parts
Part one runs in reasonable time. Part two takes an eternity, nothing to be proud of. Now that I know the logic is sound I may have a go at optimizing ir later. One obvious speedup to part two is to pre calculate the total release given a starting position and remaining time, calculate the release available from any valve. The table would not be huge but would turn a calculation into a lookup with the advantage that if there is no value stored there is insufficient remaining time