Update december_16.adb
Got the execution time down to six hours. Generated a three dimensional table of the total release that can be achieved , given a current room, destination valve and the remaining time. Also deals with three cases elephant has more remaining time than the human, human has more remaining time than the elephant and equal times. Where the times are not equal, only the operator with the more remaining time selects a valve, if there is no valve left for them to operate only then does the other operator operate a valve, if possible. The next step is to generated a two dimensional table of the set of valves that can be operated given a current room and the remaining. The intersection of that set and the set of valves still to be operated is the set of destinations that need to be examined. This should significantly reduce the search space!